Start resolving container naming and parent tracking issue
Add a new tool `render-recipe` to replace `j2render`. This allows including
the Containerfile content in the rendered recipes and implements simple
lineage tracking. Currently, the container names are used for parent
tracking, which is unreliable and makes the names awkward.

Having the Containerfile content included in the rendered recipe makes
the recipe self-sufficient and eliminates the need for additional files
to produce the images, although we will still save the Containerfiles
for reference. The lineage tracking adds checksums of the rendered
containerfiles for each image and tracks its ancestry, making it possible
to detect when a rebuild is needed if an ancestor changes.

The `build-images` script can use the output of the new tool, as it only
adds fields. However, `build-images`  but does not yet use the included
Containerfile or lineage tracking for anything.